Change buckets from pointers to indexing
This seems to be a bit faster in my tests
Also, enable & fix a few more compiler warnings.

+++ test/fuzz/insert_erase.cpp
@@ -21,7 +21,7 @@ struct DummyHash {
     using is_avalanching = void;
 
     auto operator()(uint64_t x) const noexcept -> size_t {
-        return x;
+        return static_cast<size_t>(x);
     }
 };
